innitiate getUsers early in fetch
