+++ packages/okta-adapter/src/filters/user.ts
@@ -14,39 +14,18 @@
 * limitations under the License.
 */
 import _ from 'lodash'
-import Joi from 'joi'
 import { logger } from '@salto-io/logging'
-import { applyFunctionToChangeData, createSchemeGuard, resolvePath, setPath } from '@salto-io/adapter-utils'
+import { applyFunctionToChangeData, resolvePath, setPath } from '@salto-io/adapter-utils'
 import { collections } from '@salto-io/lowerdash'
 import { Change, getChangeData, InstanceElement, isInstanceElement } from '@salto-io/adapter-api'
-import { client as clientUtils } from '@salto-io/adapter-components'
 import { FilterCreator } from '../filter'
 import { ACCESS_POLICY_RULE_TYPE_NAME, AUTHORIZATION_POLICY_RULE, GROUP_RULE_TYPE_NAME, MFA_RULE_TYPE_NAME, PASSWORD_RULE_TYPE_NAME, SIGN_ON_RULE_TYPE_NAME } from '../constants'
 import { FETCH_CONFIG } from '../config'
+import { getUsers } from '../user_utils'
 
 const log = logger(module)
-const { toArrayAsync, awu } = collections.asynciterable
+const { awu } = collections.asynciterable
 const { makeArray } = collections.array
-
-export type User = {
-  id: string
-  profile: {
-    login: string
-  }
-}
-
-const USER_SCHEMA = Joi.object({
-  id: Joi.string().required(),
-  profile: Joi.object({
-    login: Joi.string().required(),
-  }).unknown(true),
-}).unknown(true)
-
-const USERS_RESPONSE_SCHEMA = Joi.array().items(USER_SCHEMA).required()
-
-export const areUsers = createSchemeGuard<User[]>(
-  USERS_RESPONSE_SCHEMA, 'Received an invalid response for the users'
-)
 
 const EXCLUDE_USERS_PATH = ['conditions', 'people', 'users', 'exclude']
 const INCLUDE_USERS_PATH = ['conditions', 'people', 'users', 'include']
@@ -64,22 +43,6 @@ export const USER_MAPPING: Record<string, string[][]> = {
 const isRelevantInstance = (instance: InstanceElement): boolean => (
   Object.keys(USER_MAPPING).includes(instance.elemID.typeName)
 )
-
-export const getUsers = async (paginator: clientUtils.Paginator): Promise<User[]> => {
-  const paginationArgs = {
-    url: '/api/v1/users',
-    paginationField: 'after',
-    // omit credentials and other unnecessary fields from the response
-    headers: { 'Content-Type': 'application/json; okta-response=omitCredentials,omitCredentialsLinks' },
-  }
-  const users = (await toArrayAsync(
-    paginator(paginationArgs, page => makeArray(page) as clientUtils.ResponseValue[])
-  )).flat()
-  if (!areUsers(users)) {
-    return []
-  }
-  return users
-}
 
 const replaceValues = (instance: InstanceElement, mapping: Record<string, string>): void => {
   const paths = USER_MAPPING[instance.elemID.typeName]
@@ -118,7 +81,7 @@ export const replaceValuesForChanges = async (
 /**
  * Replaces user ids with login name, when 'convertUsersIds' config flag is enabled
  */
-const filterCreator: FilterCreator = ({ paginator, config }) => {
+const filterCreator: FilterCreator = ({ paginator, config, usersPromise }) => {
   let userIdToLogin: Record<string, string> = {}
   return {
     name: 'usersFilter',
@@ -127,8 +90,8 @@ const filterCreator: FilterCreator = ({ paginator, config }) => {
         log.debug('Converting user ids was disabled (onFetch)')
         return
       }
-      const users = await getUsers(paginator)
-      if (_.isEmpty(users)) {
+      const users = await usersPromise
+      if (!users || _.isEmpty(users)) {
         log.warn('Could not find any users (onFetch)')
         return
       }
