+++ packages/okta-adapter/test/filters/user.test.ts
@@ -18,7 +18,7 @@ import { client as clientUtils, filterUtils } from '@salto-io/adapter-components
 import { mockFunction } from '@salto-io/test-utils'
 import { DEFAULT_CONFIG, FETCH_CONFIG } from '../../src/config'
 import { ACCESS_POLICY_RULE_TYPE_NAME, GROUP_RULE_TYPE_NAME, OKTA } from '../../src/constants'
-import userFilter, { getUsers } from '../../src/filters/user'
+import userFilter from '../../src/filters/user'
 import { getFilterParams } from '../utils'
 
 describe('user filter', () => {
@@ -245,41 +245,6 @@ describe('user filter', () => {
           people: { users: { exclude: ['b@a.com'], include: ['a@a.com', 'c@a.com', 'd@a.com'] } },
         },
       })
-    })
-  })
-  describe('getUsers', () => {
-    const mockPaginator = mockFunction<clientUtils.Paginator>()
-      .mockImplementationOnce(async function *get() {
-        yield [
-          { id: '111', profile: { login: 'a@a.com', name: 'a' } },
-          { id: '222', profile: { login: 'b@a.com' } },
-        ]
-      })
-      .mockImplementationOnce(async function *get() {
-        yield [
-          { id: '111', profile: { name: 'a' } },
-          { id: '222', profile: { name: 'b' } },
-        ]
-      })
-    it('it should return a list of users', async () => {
-      const users = await getUsers(mockPaginator)
-      expect(users).toEqual([
-        { id: '111', profile: { login: 'a@a.com', name: 'a' } },
-        { id: '222', profile: { login: 'b@a.com' } },
-      ])
-      expect(mockPaginator).toHaveBeenNthCalledWith(
-        1,
-        {
-          url: '/api/v1/users',
-          headers: { 'Content-Type': 'application/json; okta-response=omitCredentials,omitCredentialsLinks' },
-          paginationField: 'after',
-        },
-        expect.anything()
-      )
-    })
-    it('it should return an empty list if response is invalid', async () => {
-      const users = await getUsers(mockPaginator)
-      expect(users).toEqual([])
     })
   })
 })
